frontend: move wallet connection to home view
The `RequestDialog` component should not care how the user connects with
a wallet. Its gets the connection via the store and that's it. It just
focuses 100% on building a transfer and executing it.

Later on, the wallet connection will move into a new component. But that
needs the new design of multiple wallets the user can select from which
will follow in future.
